converted NewScore10InstrPutNotesSaveClose.js to TC
